requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pymongo==4.6.0
joblib==1.3.2
scipy==1.11.4
